- Made controller variables public - Made PID variables public - Added debug pwm pin to motors - Changed type pid params to floats - Added timeout on i2c reading for mpu driver

+++ include/motor_mixer.h
@@ -5,10 +5,10 @@
 #include "stdint.h"
 
 typedef struct {
-    uint16_t m1;
-    uint16_t m2;
-    uint16_t m3;
-    uint16_t m4;
+    float m1;
+    float m2;
+    float m3;
+    float m4;
 }__attribute__((packed)) motor_mixer_command_t;
